requirements.txt
click==8.1.8
colorama==0.4.6
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.0.1
SQLAlchemy==1.4.49
typing_extensions==4.12.2
Werkzeug==2.2.2
